+++ packages/docs/src/components/docs/index.ts
@@ -1,3 +1,6 @@
 export * from './BackToTop'
 export * from './Card'
 export * from './CookiesBanner'
+export * from './CookiesConsent'
+export * from './CreditCardNumber'
+export * from './Footer'